requirements.txt
PyMuPDF==1.24.6
python-docx
pandas
scikit-learn
rake-nltk
nltk
langchain-core
langchain-community
bayesian-optimization
fitz
transformers
torch
pdf2image
tqdm
Pillow
accelerate
frontend
tabula-py
JPype1
torch
torchvision
torchaudio